2.13 upgrade (#841)
Gradle to SBT Replacement, Scala 2.13 & Kafka 2.8.1 and 3.1.0 Upgrade
* upgrading build to SBT, including replacing build shadowJar with assembly plugin
* upgrading to scala 2.13 involving many small changes to fix linting and ensure appropriate collections conversions
* building multiple versions of each of the connectors for both Kafka 2.x and 3.x to fix compatibility issues with libraries, using an sbt project matrix officially supporting latest Kafka 2.8.1 and Kafka 3.1.0
* improving the build time by running in parallel
* improving some of the flakiest tests
==============================
* Remove gradle
* Add sbt
* Common dependencies
* Common module complete and avro4s tests working again
* S3 Module fully working
* Coap connector
* Cassandra module
* Use docker version of s3proxy to allow running tests in parallel
* use testcontainers to allow dynamic ports and reduce conflicting dependencies.  Lots of dependency fun.  Project will require fixed jdk 1.8
* Embedded ES is deprecated, so switching to use same approach as ES7 module, fixes tests.
* Experiment building multiple versions for Kafka 2.8 and 3.1
* Fix test reports for modules
* Upload artifacts test
* configure to run junit tests
* align module and artifact names
* update connectorPath method
* output jars to libs subfolder
* e2e fixes
* update pipeline to run e2e tests
* upload and download assemblies for e2e test
* wildcard target path for jar
* store artifact by name
* update cache target key
* delete artifact on failure
* remove target cacheing
* assign unique name to assembly artifact
* switch to different github action to delete artifact
* split out jobs in preparation for it test stage
* test and build assembly in parallel
* Post-merge fixes

+++ kafka-connect-aws-s3/src/main/scala/io/lenses/streamreactor/connect/aws/s3/config/S3FlushSettings.scala
@@ -54,7 +54,7 @@ trait S3FlushSettings extends BaseSettings {
   }
 
   private def flushSize(kcql: Kcql): Option[Long] = {
-    Option(kcql.getWithFlushSize).filter(_ > 0).orElse(Some(defaultFlushSize))
+    Option(kcql.getWithFlushSize).filter(_ > 0).orElse(Some(defaultFlushSize.toLong))
   }
 
   private def flushCount(kcql: Kcql): Option[Long] = {
