+++ kafka-connect-cassandra/src/main/scala/com/datamountaineer/streamreactor/connect/cassandra/utils/CassandraUtils.scala
@@ -20,13 +20,12 @@ import java.time.{Instant, ZoneId}
 import java.time.format.DateTimeFormatter
 import java.time.temporal.ChronoUnit
 import java.util
-
 import com.datamountaineer.kcql.Kcql
 import com.datamountaineer.streamreactor.connect.cassandra.config.BucketMode.{BucketMode, DAY, HOUR, MINUTE, SECOND}
 import com.datastax.driver.core.Cluster
 import org.apache.kafka.connect.errors.ConnectException
 
-import scala.collection.JavaConverters._
+import scala.jdk.CollectionConverters.CollectionHasAsScala
 
 /**
   * Created by andrew@datamountaineer.com on 21/04/16.
@@ -80,7 +79,7 @@ object CassandraUtils {
     dates.add(formatter.format(previousDate))
 
     if (difference > 0) {
-      for (f <- 1 to difference.toInt) {
+      for (f <- 1L to difference) {
         dates.add(formatter.format(previousDate.plus(f, unit)))
       }
     }
