+++ kafka-connect-redis/src/main/scala/com/datamountaineer/streamreactor/connect/redis/sink/writer/RedisInsertSortedSet.scala
@@ -24,7 +24,7 @@ import com.datamountaineer.streamreactor.connect.json.SimpleJsonConverter
 import com.datamountaineer.streamreactor.connect.redis.sink.config.{RedisKCQLSetting, RedisSinkSettings}
 import org.apache.kafka.connect.sink.SinkRecord
 
-import scala.collection.JavaConverters._
+import scala.jdk.CollectionConverters.ListHasAsScala
 import scala.util.Try
 
 /**
@@ -88,7 +88,7 @@ class RedisInsertSortedSet(sinkSettings: RedisSinkSettings) extends RedisWriter 
               if (response == 1) {
                 val ttl = KCQL.kcqlConfig.getTTL
                 if (ttl > 0) {
-                  jedis.expire(sortedSetName, ttl.toInt)
+                  jedis.expire(sortedSetName, ttl)
                 }
               }
             }
